chore: move cli runner to `reth-node-core`
