+++ bin/reth/src/commands/debug_cmd/build_block.rs
@@ -5,8 +5,8 @@ use crate::{
         utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
         DatabaseArgs,
     },
+    core::cli::runner::CliContext,
     dirs::{DataDirPath, MaybePlatformPath},
-    runner::CliContext,
 };
 use alloy_rlp::Decodable;
 use clap::Parser;
