+++ bin/reth/src/commands/debug_cmd/execution.rs
@@ -6,8 +6,8 @@ use crate::{
         utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
         DatabaseArgs, NetworkArgs,
     },
+    core::cli::runner::CliContext,
     dirs::{DataDirPath, MaybePlatformPath},
-    runner::CliContext,
     utils::get_single_header,
 };
 use clap::Parser;
