+++ bin/reth/src/commands/debug_cmd/mod.rs
@@ -1,8 +1,7 @@
 //! `reth debug` command. Collection of various debugging routines.
 
+use crate::core::cli::runner::CliContext;
 use clap::{Parser, Subcommand};
-
-use crate::runner::CliContext;
 
 mod build_block;
 pub mod engine_api_store;
