+++ bin/reth/src/commands/recover/storage_tries.rs
@@ -1,7 +1,7 @@
 use crate::{
     args::utils::{chain_help, genesis_value_parser, SUPPORTED_CHAINS},
+    core::cli::runner::CliContext,
     dirs::{DataDirPath, MaybePlatformPath},
-    runner::CliContext,
 };
 use clap::Parser;
 use reth_db::{
