+++ bin/reth/src/lib.rs
@@ -29,7 +29,6 @@
 pub mod builder;
 pub mod cli;
 pub mod commands;
-pub mod runner;
 pub mod utils;
 
 /// Re-exported payload related types
