+++ crates/node-core/Cargo.toml
@@ -45,8 +45,10 @@ reth-prune.workspace = true
 reth-blockchain-tree.workspace = true
 revm-inspectors.workspace = true
 reth-snapshot.workspace = true
-reth-optimism-payload-builder = { workspace = true, optional = true }
 reth-eth-wire.workspace = true
+
+# `optimism` feature
+reth-optimism-payload-builder = { workspace = true, optional = true }
 
 # async
 tokio.workspace = true
