+++ crates/node-core/src/cli/mod.rs
@@ -4,3 +4,4 @@ pub mod components;
 pub mod config;
 pub mod db_type;
 pub mod ext;
+pub mod runner;
